userlib: add sys_send2 facade fn
Most users of sys_send turn around and slice their reply buffer based on
the returned length. The compiler has no information about the value of
the returned length, so it inserts checks and conditional panics. But
because the kernel guarantees properties of the returned length, we can
do this more cheaply using unsafe.